Allow 0 sized ibuf_open() and mark stack buffers by setting the fd to -2.
For imsgs we want to be able to use ibufs even for empty messages to stash
the fd into those. For that adjust the ibuf code to allow that.
This adds an internal IBUF_FD_MARK_ON_STACK define that is now used
for on stack ibufs instead of setting max to 0.