Rewrite the handling of the various stop and continue signal
Only deliver the signal to one thread and stop the other threads
from there. There are some complications because threads can be already
stopped or asleep. In that case handle this inside ptsignal().

For continue the situation is a bit simpler since the process was either
stopped and all threads can be restarted or it was not. Still the continue
signal needs to be delivered to userland (again only once) if not SIG_DEF.