Shortcut cursig when called during sleep setup.
Add deep flag to function argument that is used by the sleep API but
nowhere else. Both calls to sleep_signal_check() should skip the ugly
bits of cursig().

In cursig() once it is clear a signal will be taken keep the signal on
the thread siglist and return.  There is no reason to stop or handle a
signal once it is clear it is not ignored. This also skips the ptrace
trap. Once the call makes it to userret the signal will be picked up
again and handled in a safe location.