Reap the reaper: move most of it back to exit1 where it belongs.
Unlock parts of exit1(). The reaper still calls uvm_exit but the uvm_purge
call inside exit1() does most of the heavy lifting. The reaper also still
calls uarea free and proc_free() or process_zap().
Doing this differently is too complex.

knote_processexit() remains in the reaper() for now. It needs locks don't
allow it to be called late in exit1().