Implement a simple bitmap API.
Functions include:
- set, test, clear: set, test and clear a bit in the map
- empty: check if a bitmap is empty (has no bit set).
- id_get and id_put: return the lowest free id in map, return an id to the map
- init, reset: initialize and free and init a map

The first 127 elements are put directly into struct bitmap without further
allocation. For maps with more than 127 elements external memory is allocated
in the set function. This memory is only freed by reset which must be called
before an object is removed containing a bitmap.

Add a bitmap unittest as well.